refactor: split service functions
