+++ src/components/DeviceWidget.jsx
@@ -1,4 +1,4 @@
-import { useState, useEffect, version } from "react";
+import { useState, useEffect } from "react";
 import PropTypes from "prop-types";
 import {
   Card,
